fix(ci): clippy 1.95 + ESLint cleanup
CI broke on the EWS-persistence commit with clippy 1.95 + ESLint
catching long-tail issues that local tooling didn't yet surface.

Workspace-level fixes:

  - `[workspace.lints.clippy]` opts out of the stylistic noise
    (doc_lazy_continuation, doc_overindented_list_items —
    ~50 pre-existing multi-line `///` comments don't match the
    new indentation rule), plus signature-shape lints we
    deliberately accept (too_many_arguments,
    type_complexity). Every crate's Cargo.toml gained the
    `[lints] workspace = true` opt-in needed for the inherit.

Per-site fixes:

  - sync-core/version.rs: `s.split(['-', '+'])` replaces the
    closure form (manual_char_comparison).
  - cal-core/types.rs: doc-list trailing continuation moved to
    a fresh paragraph so the bullet group closes cleanly.
  - cal-adapter-local/lib.rs: dropped the inner `//!` doc on
    `pub mod test_support` so it doesn't collide with the outer
    `///` (item-with-both-attrs).
  - sync-adapter-googledrive/lib.rs: `Err(SyncError::Auth(_))`
    pattern replaces the `if matches!(...)` guard.
  - plugin-sdk/dispatch.rs: `#[allow(clippy::not_unsafe_ptr_arg_deref)]`
    with rationale doc — the unsafe block contract is the
    point.
  - src-tauri/commands/plugins.rs: `&PathBuf` → `&Path` arg.
  - src-tauri/conflicts.rs: `#[allow(clippy::should_implement_trait)]`
    on the two `from_str` methods (they predate the trait
    consideration and the name reads well in callers).
  - sync-core derives in tests: replaced `let mut state =
    SyncedFolderState::default(); state.x = ...` with struct-
    literal `..Default::default()` form
    (field_reassign_with_default) in cal-adapter-ews + cal-
    adapter-caldav.
  - 9 test files: stripped needless `&` from
    `DbHandle::open(&dir.path().join(…))` and `to_value(&fixture_event(…))`.
  - sound_assets.rs + commands/user_prefs.rs: moved post-
    `mod tests` items above the test module
    (items_after_test_module).

ESLint:

  - src/hooks/{useDeferredLoading,useEventTabNavigation}.test.ts:
    dropped `/// <reference types="vitest" />` — the
    `import { … } from 'vitest'` line already pulls in the
    globals.

155 aperio + 136 cal-adapter-ews tests still pass. Workspace
clippy and `cargo fmt --check` both clean.

+++ crates/sync-adapter-webdav/src/lib.rs
@@ -520,7 +520,7 @@ fn percent_decode(s: &str) -> Option<String> {
             }
             let hi = (bytes[i + 1] as char).to_digit(16)?;
             let lo = (bytes[i + 2] as char).to_digit(16)?;
-            out.push(char::from_u32((hi * 16 + lo) as u32)?);
+            out.push(char::from_u32(hi * 16 + lo)?);
             i += 3;
         } else {
             out.push(b as char);
